requirements.txt
numpy
matplotlib
scikit-learn
chess
CairoSVG
opencv-python
torch
torchinfo
torchvision
torchsummary
mlflow
tqdm
seaborn
pytubefix
ipywidgets
gradio
roboflow
ultralytics
python-dotenv